// src/store/mutation-types.ts
export enum MutationTypes {
  SET_PLAYER_OPEN = "SET_PLAYER_OPEN",
  PAUSE = "PAUSE",
  PLAY_NEW = "PLAY_NEW",
  SET_PLAYING = "SET_PLAYING",
  TOGGLE_PLAYING = "TOGGLE_PLAYING",
  SEEK = "SEEK",
  NEXT = "NEXT",
  PREV = "PREV",
  FAV = "FAV",
  LOGGED_IN = "LOGGED_IN",
  LOGOUT = "LOGOUT",
  ADD_SONGS = "ADD_SONGS",
  PERCENT_PLAYED = "PERCENT_PLAYED",
  AUDIO_TIME = "AUDIO_TIME"
}